Bug 1243585 - Add short-circuits when there are no pings. r=sebastian
Ideally, we'd just not run the service but skimming, I see no clean way to do
that with the existing code.

MozReview-Commit-ID: HRU1PAmYoil

+++ mobile/android/base/java/org/mozilla/gecko/telemetry/stores/TelemetryJSONFilePingStore.java
@@ -186,6 +186,10 @@ public class TelemetryJSONFilePingStore implements TelemetryPingStore {
 
     @Override
     public void onUploadAttemptComplete(final Set<Integer> successfulRemoveIDs) {
+        if (successfulRemoveIDs.isEmpty()) {
+            return;
+        }
+
         final File[] files = storeDir.listFiles(new PingFileFilter(successfulRemoveIDs));
         for (final File file : files) {
             file.delete();
